chore: remove repeated code
